requirements.txt
requests
confluent-kafka
mysql
mysql-connector-python
PyYAML
pymysql
datadog
